requirements.txt
pypdf
langchain
langchain_community
langchain-openai
rapidocr-onnxruntime
langchainhub
langchain-chroma